Preserve existing configuration with Fly.configure
This commit changes `Fly.configure` to not discard the existing
configuration before yielding.  This prevents confusing behavior when
`Fly.configure` is called from multiple initializers or when
`Fly.configure` is called after setting a value on `Fly.configuration`.
It is also more in line with e.g. `Rails.application.configure`.